+++ testing/testing_zsymv.cpp
@@ -1,9 +1,9 @@
 /*
- *  -- MAGMA (version 1.2.0) --
+ *  -- MAGMA (version 1.2.1) --
  *     Univ. of Tennessee, Knoxville
  *     Univ. of California, Berkeley
  *     Univ. of Colorado, Denver
- *     May 2012
+ *     June 2012
  *
  *  @precisions normal z -> c
  *
@@ -12,7 +12,6 @@
 #include <stdio.h>
 #include <string.h>
 #include <math.h>
-#include <cuda.h>
 #include <cuda_runtime_api.h>
 #include <cublas.h>
 #include <cblas.h>
@@ -100,8 +99,8 @@ int main(int argc, char **argv)
         lda = ((m+31)/32)*32;
         flops = FLOPS( (double)m ) / 1e6;
 
-        printf(      "%5d ", m );
-        fprintf( fp, "%5d ", m );
+        printf(      "%5d ", (int) m );
+        fprintf( fp, "%5d ", (int) m );
 
         vecsize = m * incx;
         lapackf77_zlarnv( &ione, ISEED, &vecsize, X );
